changed size_t to int on outer loops

+++ src/convolve.cpp
@@ -138,7 +138,7 @@ Napi::Value Convolve_CPU(const Napi::CallbackInfo& info) {
     float* outData = output.Data();
 
     #pragma omp for schedule(static)
-    for (size_t f = 0; f < num_filters; f++) {
+    for (int f = 0; f < num_filters; f++) {
         float bias = biases[f];
 
         for (size_t oh = 0; oh < output_height; oh++) {
@@ -287,7 +287,7 @@ Napi::Value ConvolveDelta_CPU(const Napi::CallbackInfo& info) {
 
     // ---- Convolution ----
     #pragma omp for schedule(static)
-    for (size_t c_out = 0; c_out < C_k; c_out++) {
+    for (int c_out = 0; c_out < C_k; c_out++) {
         for (size_t h = 0; h < oH; h++) {
             for (size_t w = 0; w < oW; w++) {
                 float sum = 0.0f;
